Changed the init and emulation layer to allow the NIC driver to allocate memory
The idea is that everything (in linux LOW mem = 1to1 mapping) that is above
linux.text should not be executable. This does not work during the boot
where we are forced to include also linux.init + something else up to 0xC0200000.
There is still a problem with the x_ref_cnt after downloading 8 times a 4 MB file.